backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-telegram-bot[job-queue]==22.5
drf-yasg==1.21.7
django-cors-headers==4.3.1
social-auth-app-django==5.4.1
django-filter==24.1
pillow>=10.1.0
setuptools<81
mailersend>=2.0.0
psutil==6.0.0
